Code cosmetics
GraphComponents and GraphType put into standalone files, Form1 renamed
to MainForm
