+++ MouseTester/MouseTester/MainForm.resx
