+++ MouseTester/MouseTester/Program.cs
@@ -13,7 +13,7 @@ namespace MouseTester
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new MainForm());
         }
     }
 }
